Removes CSP headers in dev environment
The CSP headers included in the requests are causing failures with
fontawesome. This in turn is causing some functionality (like pinning or
banning specific items in the builders) to disappear.

This patch removes the CSP headers from requests if the app is launched
in dev mode. It does make the app vunlerable to XSS attacks but, hey,
it's dev mode. Don't run this in production anyway.

Fixes #507